Framework update and Automation
Had to upgrade to .NET 8.0 to get a feature working for the MGS3 Guard Modifying tool. The feature for this is just to allow the user to run MGS3 while using the tool and allow the changing of coordinates based on where the player has Snake in the game. This will be more efficient than entering in each guard coordinate by hand.

Only thing that really changed because of this was how the Download Manager called downloads, everything still works like it did before.

The Automation for downloading tools, 3D Models and Textures works great tested on 4 different devices all with 100% success rate, so can start adding in more models for users to tinker with. All hardcoded paths removed from the testing for this and it fully sets up based on the user's JSON config now.