MGS2 Guard Editor
Added in a new tool I've been working on that swaps around the guard positions in MGS2. It's not perfect atm at the parsing is kind of hit and miss past a certain point. I could potentially be parsing more than the guards. However it's good enough for a test while I work on refining it.