More Models and Cleaner 3D Model UI
Big thanks to Afevis for a giant list of all the PS2 -> MC file name changes, helped me get more models in the download to test. The Textures now don't get super blurry due to zooming in to the picturebox's max size and I moved them 30px to the left so they stopped clipping the 3D Model Renderer, the texture name now also appears on top and centered above the picturebox.