MGS1 Modding Form Implemented
Given that the only mod to use on the Master Collection for MGS1 is: https://github.com/nuggslet/MGSM2Fix

The MGS1 form is more or less the same as the MGSHDFix form that pops up in the MGS2/3 forms. Might look at adding in support for Reshades and Save files in the future though if people are interested in that.